+++ NEWS.md
@@ -20,7 +20,9 @@
     - Optimized using the same strategy as other plotting functions. Unneccessary panel text is removed and caption added ([issue #47](https://github.com/helseprofil/qualcontrol/issues/47)).
     - Strata with only 1 observations is filtered out ([issue #52](https://github.com/helseprofil/qualcontrol/issues/52))
 7. `aggregate_cube` now uses collapse, and keeps unique and max values of columns that cannot be aggregated. 
-8. Code cleaning
+8. `read_files` now collects naboprikk-columns into one 0|1-column
+9. `identify_coltypes` now collects censor-columns
+10. Code cleaning
     - Removed all use of get/mget
     - Implemented the env-argument in data.table for cleaner syntax